Make applications wait for graceful close. Fix MSS to be 16 bits rather than 32.

+++ include/tcpr/types.h
@@ -21,11 +21,11 @@ struct tcpr {
 	uint32_t ack;
 	uint32_t fin;
 	uint32_t seq;
-	uint32_t win;
+	uint16_t win;
 	struct {
 		uint32_t ack;
 		uint32_t fin;
-		uint32_t win;
+		uint16_t win;
 		uint8_t have_fin;
 		uint8_t have_ack;
 	} peer;
